renamed new_a2c to a2c
